requirements.txt
openai==1.51.2
anthropic==0.34.2
pydantic==2.9.2
toml==0.10.2
PyPDF2==3.0.1
Pillow==10.4.0
python-magic==0.4.27
scikit-learn==1.5.2
numpy==2.1.2
pandas==2.2.3
pytest==8.3.3
pytest-cov==5.0.0
black==24.8.0
flake8==7.1.1
mypy==1.11.2
click==8.1.7
rich==13.8.1
textdistance==4.6.3
python-dateutil==2.9.0